src/top/L632/SolutionTest.java
package top.L632;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

import java.util.List;

class SolutionTest {

    @Test
    public void case1() {
        Solution solution = new Solution();
        Assertions.assertArrayEquals(new int[]{20, 24}, solution.smallestRange(List.of(
                List.of(4, 10, 15, 24, 26),
                List.of(0, 9, 12, 20),
                List.of(5, 18, 22, 30)
        )));
        Assertions.assertArrayEquals(new int[]{20, 24}, solution.smallestRange2(List.of(
                List.of(4, 10, 15, 24, 26),
                List.of(0, 9, 12, 20),
                List.of(5, 18, 22, 30)
        )));
    }

    @Test
    public void case2() {
        Solution solution = new Solution();
        Assertions.assertArrayEquals(new int[]{1, 1}, solution.smallestRange(List.of(
                List.of(1, 2, 3),
                List.of(1, 2, 3),
                List.of(1, 2, 3)
        )));
        Assertions.assertArrayEquals(new int[]{1, 1}, solution.smallestRange2(List.of(
                List.of(1, 2, 3),
                List.of(1, 2, 3),
                List.of(1, 2, 3)
        )));
    }

}